+++ Source/Drivers/dbebr.driver.firedac.pas
@@ -75,7 +75,7 @@ type
     destructor Destroy; override;
     procedure ExecuteDirect; override;
     function ExecuteQuery: IDBResultSet; override;
-    function RowsAffected: Integer; override;
+    function RowsAffected: UInt32; override;
   end;
 
   TDriverResultSetFireDAC = class(TDriverResultSet<TFDQuery>)
@@ -91,10 +91,10 @@ type
     procedure ApplyUpdates; override;
     function NotEof: Boolean; override;
     function GetFieldValue(const AFieldName: string): Variant; overload; override;
-    function GetFieldValue(const AFieldIndex: Integer): Variant; overload; override;
+    function GetFieldValue(const AFieldIndex: Uint16): Variant; overload; override;
     function GetFieldType(const AFieldName: string): TFieldType; overload; override;
     function GetField(const AFieldName: string): TField; override;
-    function RowsAffected: Integer; override;
+    function RowsAffected: Uint32; override;
     function IsUniDirectional: Boolean; override;
     function IsReadOnly: Boolean; override;
     function IsCachedUpdates: Boolean; override;
@@ -149,7 +149,7 @@ end;
 procedure TDriverFireDAC.ExecuteDirect(const ASQL: string; const AParams: TParams);
 var
   LExeSQL: TFDQuery;
-  LFor: Integer;
+  LFor: Int16;
 begin
   LExeSQL := TFDQuery.Create(nil);
   try
@@ -205,7 +205,7 @@ end;
 procedure TDriverFireDAC.ApplyUpdates(const ADataSets: array of IDBResultSet);
 var
   LDataSets: array of TCustomDataSet;
-  LFor: Integer;
+  LFor: Int16;
 begin
   SetLength(LDataSets, Length(ADataSets));
   for LFor := Low(ADataSets) to High(ADataSets) do
@@ -270,7 +270,7 @@ end;
 function TDriverQueryFireDAC.ExecuteQuery: IDBResultSet;
 var
   LResultSet: TFDQuery;
-  LFor: Integer;
+  LFor: UInt16;
 begin
   LResultSet := TFDQuery.Create(nil);
   try
@@ -300,7 +300,7 @@ begin
   end;
 end;
 
-function TDriverQueryFireDAC.RowsAffected: Integer;
+function TDriverQueryFireDAC.RowsAffected: UInt32;
 begin
   Result := FFDQuery.RowsAffected;
 end;
@@ -357,7 +357,7 @@ begin
   Result := FDataSet.FieldByName(AFieldName).DataType;
 end;
 
-function TDriverResultSetFireDAC.GetFieldValue(const AFieldIndex: Integer): Variant;
+function TDriverResultSetFireDAC.GetFieldValue(const AFieldIndex: Uint16): Variant;
 begin
   if AFieldIndex > FDataSet.FieldCount -1  then
     Exit(Variants.Null);
@@ -392,7 +392,7 @@ begin
   Result := not FDataSet.Eof;
 end;
 
-function TDriverResultSetFireDAC.RowsAffected: Integer;
+function TDriverResultSetFireDAC.RowsAffected: UInt32;
 begin
   Result := FDataSet.RowsAffected;
 end;
