+++ Source/Drivers/dbebr.driver.odac.pas
@@ -55,7 +55,7 @@ type
     destructor Destroy; override;
     function NotEof: Boolean; override;
     function GetFieldValue(const AFieldName: string): Variant; overload; override;
-    function GetFieldValue(const AFieldIndex: Integer): Variant; overload; override;
+    function GetFieldValue(const AFieldIndex: UInt16): Variant; overload; override;
     function GetFieldType(const AFieldName: string): TFieldType; overload; override;
     function GetField(const AFieldName: string): TField; override;
   end;
@@ -123,7 +123,7 @@ procedure TDriverODAC.ExecuteDirect(const ASQL: string;
   const AParams: TParams);
 var
   LExeSQL: TOraQuery;
-  LFor: Integer;
+  LFor: UInt16;
 begin
   LExeSQL := TOraQuery.Create(nil);
 
@@ -210,7 +210,7 @@ end;
 function TDriverQueryODAC.ExecuteQuery: IDBResultSet;
 var
   LResultSet: TOraQuery;
-  LFor: Integer;
+  LFor: UInt16;
 begin
   LResultSet := TOraQuery.Create(nil);
   try
@@ -269,8 +269,7 @@ begin
   Result := FDataSet.FieldByName(AFieldName).DataType;
 end;
 
-function TDriverResultSetODAC.GetFieldValue(
-  const AFieldIndex: Integer): Variant;
+function TDriverResultSetODAC.GetFieldValue(const AFieldIndex: UInt16): Variant;
 begin
   if AFieldIndex > FDataSet.FieldCount -1  then
     Exit(Null);
