+++ Source/Drivers/dbebr.driver.sqldirect.pas
@@ -79,7 +79,7 @@ type
     destructor Destroy; override;
     function NotEof: Boolean; override;
     function GetFieldValue(const AFieldName: string): Variant; overload; override;
-    function GetFieldValue(const AFieldIndex: Integer): Variant; overload; override;
+    function GetFieldValue(const AFieldIndex: UInt16): Variant; overload; override;
     function GetFieldType(const AFieldName: string): TFieldType; overload; override;
     function GetField(const AFieldName: string): TField; override;
   end;
@@ -125,7 +125,7 @@ end;
 procedure TDriverSQLDirect.ExecuteDirect(const ASQL: string; const AParams: TParams);
 var
   LExeSQL: TSDQuery;
-  LFor: Integer;
+  LFor: UInt16;
 begin
   LExeSQL := TSDQuery.Create(nil);
   try
@@ -232,7 +232,7 @@ end;
 function TDriverQueryFireDAC.ExecuteQuery: IDBResultSet;
 var
   LResultSet: TSDQuery;
-  LFor: Integer;
+  LFor: UInt16;
 begin
   LResultSet := TSDQuery.Create(nil);
   try
@@ -301,7 +301,7 @@ begin
   Result := FDataSet.FieldByName(AFieldName).DataType;
 end;
 
-function TDriverResultSetFireDAC.GetFieldValue(const AFieldIndex: Integer): Variant;
+function TDriverResultSetFireDAC.GetFieldValue(const AFieldIndex: UInt16): Variant;
 begin
   if AFieldIndex > FDataSet.FieldCount -1  then
     Exit(Variants.Null);
